Edit Product Photo Gallery

+++ app/Livewire/Admin/AdminProductComponent.php
@@ -20,11 +20,16 @@ class AdminProductComponent extends Component
         if($product->images)
         {
             $images = explode(",", $product->images);
+
             foreach($images as $image)
             {
-                unlink('assets\images\products'.'/'.$image);
+                if($image)
+                {
+                    unlink('assets\images\products'.'/'.$image);
+                }   
             }
         }
+
         $product->delete();
         session()->flash('message','Product has been deleted successfully!');
     }
